* removing asid from all tests and moving into supporting classes
