+++ integration-tests/integration_tests/end_to_end_tests/int_summary_care_record_adaptor_tests.py
@@ -1,6 +1,5 @@
 from unittest import TestCase
 from integration_tests.helpers.build_message import build_message
-from integration_tests.helpers.methods import get_asid
 from integration_tests.http.scr_http_request_builder import ScrHttpRequestBuilder
 from integration_tests.json.json_assertor import JsonResponseAssertor
 
@@ -34,7 +33,7 @@ class ScrAdaptorTests(TestCase):
 
     def test_should_return_success_response_from_spine_as_json(self):
         # Arrange
-        scr_json, message_id = build_message('json_16UK05', get_asid(), '9689174606')
+        scr_json, message_id = build_message('json_16UK05', '9689174606')
 
         # Act
         response = ScrHttpRequestBuilder() \
